refactor recipes hooks tests
