+++ tests/test_recipes_hooks.py
@@ -1,81 +1,49 @@
 from bocadillo import API
 from bocadillo.recipes import Recipe
+from .utils import async_function_hooks
 
 
 def test_on_async_function_view(api: API):
     numbers = Recipe('numbers')
 
-    before_called = False
-    after_called = False
+    with async_function_hooks() as (before, after):
 
-    async def before(req, res, params):
-        nonlocal before_called
-        before_called = True
+        @numbers.before(before)
+        @numbers.after(after)
+        @numbers.route('/real')
+        async def real_numbers(req, res):
+            pass
 
-    async def after(req, res, params):
-        nonlocal after_called
-        after_called = True
-
-    @numbers.before(before)
-    @numbers.after(after)
-    @numbers.route('/real')
-    async def real_numbers(req, res):
-        pass
-
-    api.recipe(numbers)
-    api.client.get('/numbers/real')
-    assert before_called
-    assert after_called
+        api.recipe(numbers)
+        api.client.get('/numbers/real')
 
 
 def test_on_sync_function_view(api: API):
     numbers = Recipe('numbers')
 
-    before_called = False
-    after_called = False
+    with async_function_hooks() as (before, after):
 
-    async def before(req, res, params):
-        nonlocal before_called
-        before_called = True
+        @numbers.before(before)
+        @numbers.after(after)
+        @numbers.route('/real')
+        def real_numbers(req, res):
+            pass
 
-    async def after(req, res, params):
-        nonlocal after_called
-        after_called = True
-
-    @numbers.before(before)
-    @numbers.after(after)
-    @numbers.route('/real')
-    def real_numbers(req, res):
-        pass
-
-    api.recipe(numbers)
-    api.client.get('/numbers/real')
-    assert before_called
-    assert after_called
+        api.recipe(numbers)
+        api.client.get('/numbers/real')
 
 
 def test_on_class_based_view(api: API):
     numbers = Recipe('numbers')
 
-    before_called = False
-    after_called = False
+    with async_function_hooks() as (before, after):
 
-    async def before(req, res, params):
-        nonlocal before_called
-        before_called = True
+        @numbers.before(before)
+        @numbers.route('/real')
+        class RealNumbers:
+            @numbers.after(after)
+            async def get(self, req, res):
+                pass
 
-    async def after(req, res, params):
-        nonlocal after_called
-        after_called = True
-
-    @numbers.before(before)
-    @numbers.route('/real')
-    class RealNumbers:
-        @numbers.after(after)
-        async def get(self, req, res):
-            pass
-
-    api.recipe(numbers)
-    api.client.get('/numbers/real')
-    assert before_called
-    assert after_called
+        api.recipe(numbers)
+        api.client.get('/numbers/real')
